[AD-543] change parameter type from int8_t to int16_t

+++ src/odbc/include/ignite/odbc/type_traits.h
@@ -333,7 +333,7 @@ namespace ignite
              * @param type Binary type.
              * @return True if unsigned or non-numeric.
              */
-            bool BinaryTypeUnsigned(int8_t type);
+            bool BinaryTypeUnsigned(int16_t type);
         }
     }
 }
